test: migrate to vitest and colocate tests
