--- jest.config.cjs
@@ -1,24 +1,0 @@
-module.exports = {
-  testEnvironment: "jsdom",
-  transform: {
-    "^.+\\.(ts|tsx)$": [
-      "@swc/jest",
-      {
-        jsc: {
-          parser: {
-            syntax: "typescript",
-            tsx: true,
-          },
-          transform: {
-            react: {
-              runtime: "classic",
-            },
-          },
-        },
-        module: {
-          type: "commonjs",
-        },
-      },
-    ],
-  },
-};
