+++ package-checks/smoke.cjs
