+++ package-checks/tsconfig.cjs.json
@@ -8,5 +8,5 @@
     "noEmit": true,
     "types": ["node", "react"]
   },
-  "files": ["package-types.mts"]
+  "files": ["./types.cts"]
 }
