+++ package.json
@@ -36,34 +36,33 @@
     }
   },
   "scripts": {
-    "test": "jest",
+    "test": "vitest run",
+    "test:watch": "vitest watch",
     "check": "npm run check:format && npm run check:lint && npm run typecheck",
     "check:format": "oxfmt --check",
     "check:lint": "oxlint",
     "lint": "npm run check:lint",
     "typecheck": "tsc --project tsconfig.json",
-    "test:package": "node __test__/package-smoke.cjs && tsc --project __test__/tsconfig.package.json --noEmit && tsc --project __test__/tsconfig.package-cjs.json --noEmit",
-    "build": "tsdown && npm run test:package",
+    "check:package": "node package-checks/smoke.cjs && tsc --project package-checks/tsconfig.esm.json --noEmit && tsc --project package-checks/tsconfig.cjs.json --noEmit",
+    "test:package": "npm run check:package",
+    "build": "tsdown && npm run check:package",
     "fix": "npm run fix:format && npm run fix:lint",
     "fix:format": "oxfmt",
     "fix:lint": "oxlint --fix",
     "prettify": "npm run fix:format"
   },
   "devDependencies": {
-    "@swc/core": "1.15.33",
-    "@swc/jest": "0.2.39",
     "@testing-library/react": "13.3.0",
-    "@types/jest": "^30.0.0",
     "@types/react": "^18.3.27",
     "@types/react-dom": "^18.3.7",
-    "jest": "28.1.2",
-    "jest-environment-jsdom": "28.1.2",
+    "jsdom": "29.1.1",
     "oxfmt": "0.50.0",
     "oxlint": "1.65.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "tsdown": "0.22.0",
-    "typescript": "^5.9.3"
+    "typescript": "^5.9.3",
+    "vitest": "4.1.7"
   },
   "peerDependencies": {
     "react": ">=17.0.2",
